#1750 Moved lbound etc test from reference_test to operation_test.
