+++ src/psyclone/tests/psyir/nodes/reference_test.py
@@ -145,38 +145,6 @@ def test_reference_accesses():
     assert (str(var_access_info)) == "test: READ"
 
 
-@pytest.mark.parametrize("operator", ["lbound", "ubound", "size"])
-def test_reference_accesses_bounds(operator, fortran_reader):
-    '''Test that the reference_accesses method behaves as expected when
-    the reference is the first argument to either the lbound or ubound
-    intrinsic as that is simply looking up the array bounds (therefore
-    var_access_info should be empty) and when the reference is the
-    second argument of either the lbound or ubound intrinsic (in which
-    case the access should be a read).
-
-    '''
-    code = f'''module test
-        contains
-        subroutine tmp()
-          real, dimension(:,:), allocatable:: a, b
-          integer :: n
-          n = {operator}(a, b(1,1))
-        end subroutine tmp
-        end module test'''
-    psyir = fortran_reader.psyir_from_source(code)
-    schedule = psyir.children[0].children[0]
-
-    # By default, the access to 'a' should not be reported as read,
-    # but the access to b must be reported:
-    vai = VariablesAccessInfo(schedule[0])
-    assert str(vai) == "b: READ, n: WRITE"
-
-    # When explicitly requested, the access to 'a' should be reported:
-    vai = VariablesAccessInfo(schedule[0],
-                              options={"COLLECT-ARRAY-SHAPE-READS": True})
-    assert str(vai) == "a: READ, b: READ, n: WRITE"
-
-
 def test_reference_can_be_copied():
     ''' Test that a reference can be copied. '''
